src/requirements.txt
blinker==1.9.0 ; python_version >= "3.12" and python_version < "4"
cachetools==5.5.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.0 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4"
cloudevents==1.11.0 ; python_version >= "3.12" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4" and platform_system == "Windows"
deprecation==2.1.0 ; python_version >= "3.12" and python_version < "4"
flask==3.1.0 ; python_version >= "3.12" and python_version < "4"
functions-framework==3.8.2 ; python_version >= "3.12" and python_version < "4"
google-api-core==2.23.0 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.36.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-storage==2.18.2 ; python_version >= "3.12" and python_version < "4.0"
google-crc32c==1.6.0 ; python_version >= "3.12" and python_version < "4.0"
google-resumable-media==2.7.2 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.66.0 ; python_version >= "3.12" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.12" and python_version < "4" and platform_system != "Windows"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.12" and python_version < "4"
jinja2==3.1.4 ; python_version >= "3.12" and python_version < "4"
markupsafe==3.0.2 ; python_version >= "3.12" and python_version < "4"
packaging==24.2 ; python_version >= "3.12" and python_version < "4"
proto-plus==1.25.0 ; python_version >= "3.12" and python_version < "4.0"
protobuf==5.29.0 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
redis==5.2.0 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
spotipy==2.24.0 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
watchdog==6.0.0 ; python_version >= "3.12" and python_version < "4"
werkzeug==3.1.3 ; python_version >= "3.12" and python_version < "4"
yt-dlp==2024.11.18 ; python_version >= "3.12" and python_version < "4.0"
ytmusicapi==1.8.2 ; python_version >= "3.12" and python_version < "4.0"
